requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.11.0
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
python-jose>=3.3.0  # voor JWT token handling
cryptography>=3.4.7  # voor veilige sleutel opslag
loguru>=0.5.3  # voor verbeterde logging
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
fastapi>=0.68.0
uvicorn>=0.15.0  # voor API server 